feat: add authentication

+++ src/pages/api/instances/[id]/index.ts
@@ -4,6 +4,8 @@ import { IMessage } from "@/types/chat";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { prisma } from "../../../../lib";
+import { getSession } from "next-auth/react";
+import { getUserByEmail } from "@/utils/queries";
 
 type IResponse = InstanceResponse | EmptyErrorResponse;
 
@@ -21,7 +23,15 @@ export default async function handler(
   });
   if (!obj) return res.status(404).send({});
 
+  const session = await getSession({ req });
+  const user = session?.user?.email
+    ? await getUserByEmail(session.user.email)
+    : null;
+
+  if (!obj.public && obj.userId !== user?.id) return res.status(401).send({});
+
   if (req.method === "DELETE") {
+    if (obj.userId !== user?.id) return res.status(401).json({});
     await prisma.instance.delete({
       where: {
         id,
@@ -34,7 +44,9 @@ export default async function handler(
   if (req.method === "GET") {
     return res.status(201).json(await repository.serializeInstance(obj));
   } else if (req.method === "PATCH") {
-    const { title } = req.body;
+    if (obj.userId !== user?.id) return res.status(401).json({});
+
+    const { title, public: _public } = req.body;
 
     const updatedObj = await prisma.instance.update({
       where: {
@@ -43,6 +55,7 @@ export default async function handler(
       data: {
         ...obj,
         title: title || obj.title,
+        public: !!_public || obj.public,
       },
     });
